requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
office365-rest-python-client>=2.4.6